moved AotUPoster to AdvancedMaterials
